chore: utils directory move to core directory
